+++ Magento/Mtf/Client/Driver/Selenium/Driver.php
@@ -569,6 +569,7 @@ class Driver implements DriverInterface
         foreach ($keys as $key) {
             $this->driver->keys($key);
         }
+        $this->triggerChangeEvent($element);
     }
 
     /**